pyproject.toml
[project]
name = "dead-letter-oracle"
version = "1.0.0"
description = "Governed MCP agent for DLQ incident resolution with closed-loop reasoning and multi-factor governance"
readme = "README.md"
license = { file = "LICENSE-2.0.txt" }
authors = [
    { name="Prasad Tiruveedi", email = "tvprasad@gmail.com" },
    { name="VPL Solutions LLC", email = "contact@vplsolutions.com"}    
]
requires-python = ">=3.12"
keywords = ["mcp", "agent", "dlq", "governance", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "mcp==1.26.0",
    "openai==2.18.0",
    "anthropic==0.86.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    "anyio==4.12.1",
    "uvicorn==0.42.0",
]

[project.optional-dependencies]
test = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "ruff>=0.9",
]

[project.scripts]
dead-letter-oracle = "main:main"

# ── Build ──────────────────────────────────────────────────────────────────────

[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_server*", "agent*", "governance*", "observability*"]

# ── Pytest ─────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "strict"

# ── Ruff ───────────────────────────────────────────────────────────────────────

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

# ── Mypy ───────────────────────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
check_untyped_defs = false

# ── Bandit ─────────────────────────────────────────────────────────────────────

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]   # assert statements used intentionally in tests
